+++ lib/src/error_tracking/chunk_ids.dart
@@ -1,0 +1,1 @@
+export 'chunk_ids_io.dart' if (dart.library.js_interop) 'chunk_ids_web.dart';